+++ X_API_Master/X.Stock.Service/Model/CustomerInfo.cs
@@ -8,6 +8,6 @@ namespace X.Stock.Service.Model
 
         public string CustomerName { get; set; }
 
-        public decimal CoinAsset { get; set; }
+        public double CoinAsset { get; set; }
     }
 }
