+++ X_API_Master/X.Stock.Service/Utils/StockService.cs
@@ -38,11 +38,11 @@ namespace X.Stock.Service.Utils
                 string.Empty);
             var reg = new Regex("\\\"(.+?)\\\"");
             var groups = reg.Matches(iresult.Content);
-            decimal stockPrice, stockKm1, stockKm2;
+            double stockPrice, stockKm1, stockKm2;
             DateTime dt;
-            if (groups.Count < 52 || !decimal.TryParse(groups[27].Groups[1].Value, out stockPrice) ||
-                !decimal.TryParse(groups[29].Groups[1].Value, out stockKm1) ||
-                !decimal.TryParse(groups[31].Groups[1].Value, out stockKm2) ||
+            if (groups.Count < 52 || !double.TryParse(groups[27].Groups[1].Value, out stockPrice) ||
+                !double.TryParse(groups[29].Groups[1].Value, out stockKm1) ||
+                !double.TryParse(groups[31].Groups[1].Value, out stockKm2) ||
                 !DateTime.TryParse(groups[51].Groups[1].Value, out dt)) return result;
             result.StockCode = groups[3].Groups[1].Value;
             result.StockName = groups[4].Groups[1].Value;
@@ -68,11 +68,11 @@ namespace X.Stock.Service.Utils
             var reg = new Regex("\\\"(.+?)\\\"");
             var groups = reg.Matches(iresult.Content);
             if (groups.Count <= 0) return result;
-            decimal stockPrice = 0, stockKm1 = 0, stockKm2 = 0;
+            double stockPrice = 0, stockKm1 = 0, stockKm2 = 0;
             result.AddRange(from Match @group in groups
                 select @group.Groups[1].ToString().Split(',')
                 into array
-                where array.Length >= 6 && decimal.TryParse(array[3], out stockPrice) && decimal.TryParse(array[4].Replace("%", string.Empty), out stockKm2) && decimal.TryParse(array[5], out stockKm1)
+                                where array.Length >= 6 && double.TryParse(array[3], out stockPrice) && double.TryParse(array[4].Replace("%", string.Empty), out stockKm2) && double.TryParse(array[5], out stockKm1)
                 select new StockInfo
                 {
                     StockCode = array[1], StockName = array[2], StockPrice = stockPrice, StockKm2 = stockKm2, StockKm1 = stockKm1
@@ -80,7 +80,7 @@ namespace X.Stock.Service.Utils
             return result;
         }
 
-        public static decimal GetBenifit(decimal cost, decimal current)
+        public static double GetBenifit(double cost, double current)
         {
             return current/cost - 1;
         }
